+++ projects/packages/publicize/src/rest-api/class-social-image-generator-controller.php
@@ -50,10 +50,13 @@ class Social_Image_Generator_Controller extends Base_Controller {
 			$this->namespace,
 			'/' . $this->rest_base . '/generate-token',
 			array(
-				'methods'             => WP_REST_Server::CREATABLE,
-				'callback'            => array( $this, 'generate_preview_token' ),
-				'permission_callback' => array( $this, 'permissions_check' ),
-				'args'                => array(
+				'methods'                        => WP_REST_Server::CREATABLE,
+				'callback'                       => array( $this, 'generate_preview_token' ),
+				'permission_callback'            => array( $this, 'permissions_check' ),
+				'private_site_security_settings' => array(
+					'allow_blog_token_access' => true,
+				),
+				'args'                           => array(
 					'text'      => array(
 						'description' => __( 'The text to be used to generate the image.', 'jetpack-publicize-pkg' ),
 						'type'        => 'string',
@@ -75,11 +78,43 @@ class Social_Image_Generator_Controller extends Base_Controller {
 						'type'        => 'string',
 						'enum'        => Templates::TEMPLATES,
 					),
+					'font'      => array(
+						'description' => __( 'The font slug.', 'jetpack-publicize-pkg' ),
+						'type'        => 'string',
+					),
 				),
-				'schema'              => array(
+				'schema'                         => array(
 					'$schema' => 'http://json-schema.org/draft-04/schema#',
 					'title'   => 'publicize-sig-generate-token',
 					'type'    => 'string',
+				),
+			)
+		);
+
+		register_rest_route(
+			$this->namespace,
+			'/' . $this->rest_base . '/font-options',
+			array(
+				'methods'             => WP_REST_Server::READABLE,
+				'callback'            => array( $this, 'get_font_options' ),
+				'permission_callback' => array( $this, 'permissions_check' ),
+				'schema'              => array(
+					'$schema' => 'http://json-schema.org/draft-04/schema#',
+					'title'   => 'publicize-sig-font-options',
+					'type'    => 'array',
+					'items'   => array(
+						'type'       => 'object',
+						'properties' => array(
+							'id'    => array(
+								'type'        => 'string',
+								'description' => __( 'Unique identifier for the font.', 'jetpack-publicize-pkg' ),
+							),
+							'label' => array(
+								'type'        => 'string',
+								'description' => __( 'The font label.', 'jetpack-publicize-pkg' ),
+							),
+						),
+					),
 				),
 			)
 		);
@@ -118,8 +153,39 @@ class Social_Image_Generator_Controller extends Base_Controller {
 			SIG\fetch_token(
 				$request->get_param( 'text' ),
 				$request->get_param( 'image_url' ),
-				$request->get_param( 'template' )
+				$request->get_param( 'template' ),
+				$request->get_param( 'font' )
 			)
 		);
+	}
+
+	/**
+	 * Returns the available font options for the social image generator.
+	 *
+	 * @param WP_REST_Request $request The request object, which includes the parameters.
+	 *
+	 * @return WP_REST_Response The response containing the font options.
+	 */
+	public function get_font_options( $request ) {
+		if ( Utils::is_wpcom() ) {
+			require_lib( 'publicize/util/social-image-generator' );
+
+			$fonts = \Publicize\Social_Image_Generator\get_font_options();
+
+			$font_options = array();
+
+			foreach ( $fonts as $id => [ 'label' => $label ] ) {
+				$font_options[] = compact( 'id', 'label' );
+			}
+
+			return rest_ensure_response( $font_options );
+		}
+
+		$response = $this->proxy_request_to_wpcom_as_blog(
+			$request,
+			'font-options'
+		);
+
+		return rest_ensure_response( $response );
 	}
 }