+++ projects/packages/stats-admin/src/class-main.php
@@ -22,7 +22,7 @@ class Main {
 	/**
 	 * Stats version.
 	 */
-	const VERSION = '0.28.2';
+	const VERSION = '0.28.3';
 
 	/**
 	 * Singleton Main instance.