+++ projects/packages/plugin-deactivation/CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.3.15] - 2025-08-05
+### Changed
+- Internal updates.
+
 ## [0.3.14] - 2025-07-23
 ### Changed
 - Internal updates.
@@ -120,6 +124,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 - Added package to intercept plugin deactivation [#27081]
 
+[0.3.15]: https://github.com/Automattic/jetpack-plugin-deactivation/compare/v0.3.14...v0.3.15
 [0.3.14]: https://github.com/Automattic/jetpack-plugin-deactivation/compare/v0.3.13...v0.3.14
 [0.3.13]: https://github.com/Automattic/jetpack-plugin-deactivation/compare/v0.3.12...v0.3.13
 [0.3.12]: https://github.com/Automattic/jetpack-plugin-deactivation/compare/v0.3.11...v0.3.12